requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn

langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers

youtube-transcript-api

python-dotenv

ipykernel

streamlit